+++ commons/ava_testsuite/unrequested_chit_spammer_test/unrequested_chit_spammer_test_.go
@@ -4,9 +4,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_networks"
-	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_services"
-	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_testsuite/rpc_workflow_runner"
+	"github.com/ava-labs/avalanche-e2e-tests/commons/ava_networks"
+	"github.com/ava-labs/avalanche-e2e-tests/commons/ava_services"
+	"github.com/ava-labs/avalanche-e2e-tests/commons/ava_testsuite/rpc_workflow_runner"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
 	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 	"github.com/palantir/stacktrace"
@@ -107,7 +107,7 @@ func (test StakingNetworkUnrequestedChitSpammerTest) GetNetworkLoader() (network
 			2,
 			2,
 			map[string]string{
-				"byzantine-behavior": "chit-spammer",
+				byzantineBehavior: chitSpammerBehavior,
 			},
 		),
 		normalNodeConfigId: *ava_networks.NewTestGeckoNetworkServiceConfig(
@@ -139,11 +139,11 @@ func (test StakingNetworkUnrequestedChitSpammerTest) GetNetworkLoader() (network
 }
 
 func (test StakingNetworkUnrequestedChitSpammerTest) GetExecutionTimeout() time.Duration {
-	return 5 * time.Minute
+	// TODO drop this when the availabilityChecker doesn't have a sleep, because we spin up a *bunch* of byzantine
+	// nodes during test execution
+	return 10 * time.Minute
 }
 
 func (test StakingNetworkUnrequestedChitSpammerTest) GetSetupBuffer() time.Duration {
-	// TODO drop this when the availabilityChecker doesn't have a sleep, because we spin up a *bunch* of nodes before test
-	//  execution starts
-	return 12 * time.Minute
+	return 4 * time.Minute
 }