+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Switch configuration IDs to strings instead of ints
 * Bump kurtosis version to get cleanup on ctrl-c
 * Pull in Kurtosis version that will print test outputs they finish, rather than waiting for all tests to finish
+* Point CI to stable Docker image based off of Denali for both gecko and gecko-byzantine
 
 # 0.7.0
 * Split `staking_network_tests` into separate files per test